% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic}
\alias{basic}
\title{basic}
\format{A data frame with 50 rows and 22 variables:
\describe{
  \item{\code{sampling}}{character. sampling instrument.}
  \item{\code{haul}}{character. number of haul.}
  \item{\code{catchday}}{integer. day of catch.}
  \item{\code{deathday}}{integer. day of death.}
  \item{\code{catch_date}}{character. date of catch.}
  \item{\code{death_date}}{character. date of death.}
  \item{\code{tl_mm}}{integer. total length fish in mm.}
  \item{\code{sl_mm}}{integer. standard length fish in mm.}
  \item{\code{wet_weight_g}}{double. wet weight fish in g.}
  \item{\code{dry_weight.schale}}{double. dry weight fish + dish in g.}
  \item{\code{weight_schale}}{double. weight dish in g.}
  \item{\code{dry_weight_g}}{double. dry weight fish in g.}
  \item{\code{oto_1.2}}{integer. first or second otolith of fish x.}
  \item{\code{oto_weight_mg}}{double. otolith weight in mg.}
  \item{\code{oto_diameter_mm}}{double. otolith diameter in mm.}
  \item{\code{foto}}{character. are fotos aviable? y = yes, n = no.}
  \item{\code{quality}}{double. quality of the fotos. 1 = good, 4 = bad.}
  \item{\code{age}}{double. age of the fish in days.}
  \item{\code{first_change}}{integer. first change in increment width at ring x.}
  \item{\code{sec_change_}}{integer. second change in increment width at ring y.}
  \item{\code{comment}}{character. comments.}
}}
\usage{
basic
}
\description{
some basic data for all fish
}
\keyword{datasets}
