% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_growth.R
\name{calc_mean_growth}
\alias{calc_mean_growth}
\title{calc_mean_growth}
\usage{
calc_mean_growth(x, y)
}
\arguments{
\item{x}{variable to group y}

\item{y}{vector/ list of vectors with numeric values the same length as x}
}
\description{
Calculate the mean growth of n fish.
}
\details{
Calculate the mean growth of n fish for each julian day. Returns mean and standard deviation.
}
\examples{
#rm(list = ls())
x <- master_tib$julday
y <- master_tib$ring_width
calc_mean_growth(x, y)

}
