% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_func.R
\name{comb_func}
\alias{comb_func}
\title{Comb_func}
\usage{
comb_func(x, y, by = "id")
}
\arguments{
\item{x}{a mastertable given by oto_makro}

\item{y}{a dfr with information you want to join to the mastertable}

\item{by}{by which vector should the dfrs be joined, default is id (must be present in both dfrs!)}
}
\description{
Plotting ring width
}
\details{
Plots ring width ~ julian day, should not include more than 12 fish and works with the output
from master_tab
}
\examples{
#rm(list = ls())
x <- oto_makro(x = example[,3], y = example[,4], fish_no = "example")
x$id <- as.character(x$id)
y <- gathering
combi <- comb_func(x = x, y = y)

}

