% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{dat}
\format{A data frame with 50 rows and 28 variables:
\describe{
  \item{\code{id}}{character. id.}
  \item{\code{ring_no}}{list. number of rings.}
  \item{\code{ring_width}}{list. ring width in µm.}
  \item{\code{hc}}{double. hatchcheck in µm.}
  \item{\code{oto_rad}}{double. radius of the otolith in µm.}
  \item{\code{age}}{double. age of the fish in days.}
  \item{\code{sampling}}{character. sampling instrument.}
  \item{\code{haul}}{character. number of haul.}
  \item{\code{catchday}}{integer. day of catch.}
  \item{\code{deathday}}{integer. day of death.}
  \item{\code{catch_date}}{character. date of catch.}
  \item{\code{death_date}}{character. date of death.}
  \item{\code{tl_mm}}{integer. total length fish in mm.}
  \item{\code{sl_mm}}{integer. standard length fish in mm.}
  \item{\code{wet_weight_g}}{double. wet weight fish in g.}
  \item{\code{dry_weight.schale}}{double. dry weight fish + dish in g.}
  \item{\code{weight_schale}}{double. weight dish in g.}
  \item{\code{dry_weight_g}}{double. dry weight fish in g.}
  \item{\code{oto_1.2}}{integer. first or second otolith of fish x.}
  \item{\code{oto_weight_mg}}{double. otolith weight in mg.}
  \item{\code{oto_diameter_mm}}{double. otolith diameter in mm.}
  \item{\code{foto}}{character. are fotos aviable? y = yes, n = no.}
  \item{\code{quality}}{double. quality of the fotos. 1 = good, 4 = bad.}
  \item{\code{first_change}}{integer. first change in increment width at ring x.}
  \item{\code{sec_change_}}{integer. second change in increment width at ring y.}
  \item{\code{comment}}{character. comments.}
  \item{\code{julday}}{list. julian day of increment formation.}
  \item{\code{move_ave}}{list. see huessy 2003.}
  \item{\code{cum_oto_rad}}{list. otolith radius at increment x.}
}}
\usage{
dat
}
\description{
data generated by myself
}
\keyword{datasets}
