% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ro}
\alias{dat_ro}
\title{dat_ro}
\format{A data frame with 99 rows and 20 variables:
\describe{
  \item{\code{catchday}}{double. day of catch.}
  \item{\code{kw}}{double. calendar week.}
  \item{\code{notes_original}}{character. notes.}
  \item{\code{location}}{character. location of catch event.}
  \item{\code{ices_rectangle}}{character. ices rectangle.}
  \item{\code{sd}}{double. ices subdivision.}
  \item{\code{slice_no}}{double. no idea.}
  \item{\code{id}}{double. id.}
  \item{\code{tl_cm}}{double. total length fish in cm.}
  \item{\code{weight_vmk_g}}{double. weight fish in g.}
  \item{\code{weight_amk_g}}{double. weight fish in g.}
  \item{\code{sex}}{double. sex. 1 = female, 2 = male.}
  \item{\code{maturity}}{double. maturity. for code ask klinger.}
  \item{\code{gonad_weight_g}}{character. gonad weight in g.}
  \item{\code{liver_Weight_g}}{character. liver weight in g.}
  \item{\code{age_year}}{character. age in years.}
  \item{\code{outer_zone}}{character. outer zone hyalin or opak? 2 = opak i guess.}
  \item{\code{readable}}{character. quality of the otolith.}
  \item{\code{note_oto_right}}{logical. notes.}
  \item{\code{note_oto_left}}{logical. notes.}
}}
\usage{
dat_ro
}
\description{
data from rostock
}
\keyword{datasets}
