% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{example}
\format{A data frame with 123 rows and 18 variables:
\describe{
  \item{\code{X}}{integer. increasing numbers from 1 on.}
  \item{\code{Features}}{integer. same as X.}
  \item{\code{Center.X.Pos.}}{integer. x coordinates on the foto.}
  \item{\code{Center.Y.Pos.}}{integer. y coordinates on the foto.}
  \item{\code{Area}}{integer. not relevant.}
  \item{\code{Length}}{integer. not relevant.}
  \item{\code{Radius}}{integer. not relevant.}
  \item{\code{Start.X.Pos.}}{integer. not relevant.}
  \item{\code{Start.Y.Pos.}}{integer. not relevant.}
  \item{\code{End.X.Pos.}}{integer. not relevant.}
  \item{\code{End.Y.Pos.}}{integer. not relevant.}
  \item{\code{Angle}}{integer. not relevant.}
  \item{\code{Avg..Dist.}}{integer. not relevant.}
  \item{\code{Min..Dist.}}{integer. not relevant.}
  \item{\code{Max..Dist.}}{integer. not relevant.}
  \item{\code{Ctr.to.Ctr.Dist.}}{integer. not relevant.}
  \item{\code{Perp..Dist.}}{integer. not relevant.}
  \item{\code{Count}}{integer. not relevant.}
}}
\usage{
example
}
\description{
exel output generated by imagePRO for fish 708
}
\keyword{datasets}
