% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scalebar.R
\name{geom_scalebar}
\alias{geom_scalebar}
\title{geom_scalebar}
\usage{
geom_scalebar(range, y, segments, size)
}
\arguments{
\item{range}{A vector with two values. Giving the limits in x direction.
It doesn't matter which value comes first (if it's not important which colour the first segment has).}

\item{y}{Value for vertical positioning.}

\item{segments}{Specifies the number of segments between lower and upper range.}

\item{size}{Specifies the width of the scale bar.}
}
\description{
Creates a scale bar in a ggplot
}
\details{
Uses geom_segment to create a black and white scale bar in horizontal direction.
Is not able to give any information about the magnification itself!
The exact length of the scale bar has to be known.
}
\examples{
#rm(list = ls())
ggplot2::ggplot(data = NULL, ggplot2::aes(x = c(2,5), y = c(3, 1))) +
   ggplot2::geom_point() +
   geom_scalebar(range = c(2.5, 4.5), y = 2, segments = 4, size = 2)

ggplot2::ggplot(data = NULL, ggplot2::aes(x = c(2,5), y = c(3, 1))) +
   ggplot2::geom_point() +
   geom_scalebar(range = c(4.5, 2.5), y = 2, segments = 4, size = 2)

}
