% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fish.R
\name{load_fish}
\alias{load_fish}
\title{load_fish}
\usage{
load_fish(fish, dir = getwd(), format = "txt", sep = "\\t")
}
\arguments{
\item{fish}{name of the file to be loaded}

\item{dir}{path to a folder}

\item{format}{format of the file to be loaded (either txt or csv, the default is txt)}

\item{sep}{separator used by the file to be loaded (default is tab)}
}
\description{
Load in the data correctly for oto_makro.
}
\details{
Automatically load in the data the way it's needed for oto_makro. Returns a R object.
}
\examples{
#rm(list = ls())
fish <- load_fish(fish = "example", format = "csv", sep = ";", dir = "F:/ototools")
str(fish)

}
