% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fish2.R
\name{load_fish2}
\alias{load_fish2}
\title{load_fish2}
\usage{
load_fish2(fish, section, dir = getwd(), sep = "\\t")
}
\arguments{
\item{fish}{name of the file to be loaded}

\item{section}{which sections to be loaded. Has to be part of the filename!}

\item{dir}{path to a folder}

\item{sep}{separator used by the file to be loaded (default is tab)}
}
\description{
Load in the data correctly for oto_makro.
}
\details{
Automatically load in the data the way it's needed for oto_makro. Returns a R object.
Instead loading only one file, load_fish2 can load multiple files for the same individual.
}
\examples{
#rm(list = ls())

}
