% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makro_once.R
\name{makro_once}
\alias{makro_once}
\title{Makro_once}
\usage{
makro_once(fish, format, sep)
}
\arguments{
\item{fish}{name of the file to be loaded}

\item{format}{format of the file to be loaded (either txt or csv)}

\item{sep}{the separator used by the file to be loaded}
}
\description{
Run some functions of ototools
}
\details{
Uses load_fish + oto_makro + plot_growth. Returns a dfr.
}
\examples{
#rm(list = ls())
fish <- "example"
format <- "csv"
x <- makro_once(fish = fish, format = format, sep = ";")
head(x)
}

