% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{makro_output}
\alias{makro_output}
\title{makro_output}
\format{A data frame with 1 rows and 6 variables:
\describe{
  \item{\code{id}}{character. id of the fish.}
  \item{\code{ring_no}}{list. number of ring.}
  \item{\code{ring_width}}{list. respective ring width.}
  \item{\code{hc}}{double. width of the hatch check.}
  \item{\code{oto_rad}}{double. length of the otolith radius.}
  \item{\code{age}}{double. estimated age of the fish.}
}}
\usage{
makro_output
}
\description{
output generated by oto_makro
}
\keyword{datasets}
