% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_func.R
\name{master_func}
\alias{master_func}
\title{Master function}
\usage{
master_func(fish, format, sep, basic, catch_day)
}
\arguments{
\item{fish}{name of the file(s) to be loaded}

\item{format}{format of the file(s) to be loaded (either txt or csv);
has to be the same for every file}

\item{sep}{the separator used by the file(s) to be loaded;
has to be the same for every file}

\item{basic}{a dfr with some basic values for the fish}

\item{catch_day}{name of the column containing the catch day (julian)}
}
\description{
Run some func of ototools
}
\details{
Uses load_fish + oto_makro + master_tab
}
\examples{
#rm(list = ls())
fish <- "example"
format <- "csv"
x <- master_func(fish = fish, format = format, sep = ";",
                 basic = gathering, catch_day = "catch_date")

}

