% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_ave.R
\name{move_ave}
\alias{move_ave}
\title{Move_ave}
\usage{
move_ave(ring_no, ring_width, data)
}
\arguments{
\item{ring_no}{colname of the column with ring numbers}

\item{ring_width}{colname of the column with ring widths}

\item{data}{dfr with both ring_no and ring_width}
}
\description{
Calculat the difference in moving averages following Huessy et al. 2003
}
\details{
Calculating the difference of the moving average from the last 10 increments and the next
10 increments. Returns absolut values.
}
\examples{
#rm(list = ls())
data <- oto_makro(x = example[,3], y = example[,4], fish_no = 1)
move_ave("ring_no", "ring_width", data)
}

