% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_makro.R
\name{oto_makro}
\alias{oto_makro}
\title{Otolith reading Makro}
\usage{
oto_makro(x, y, fish_no, sep = ";")
}
\arguments{
\item{x}{output from imagePRO (x coordinate)}

\item{y}{output from imagePRO (y coordinate)}

\item{fish_no}{number of the fish or whatever chosen to identify them}

\item{sep}{the seperator for the file to be created (default is ";")}
}
\description{
Package to transform coordinates from imagePRO into distances
}
\details{
Takes coordinates delivered by imagePRO and calculates age, hatch check width and ring
width for the otolith
}
\examples{
#rm(list = ls())
#returns a data.frame with ring_width and more
x <- oto_makro(x = example[,3], y = example[,4], fish_no = 1)
head(x)

}

