% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geom_point.R
\name{plot_geom_point}
\alias{plot_geom_point}
\title{Plot function}
\usage{
plot_geom_point(data, x_var, y_var)
}
\arguments{
\item{data}{a dfr holding x_var and y_var}

\item{x_var}{x variable for a geom_point plot}

\item{y_var}{y variable for a geom_point plot}
}
\description{
Plotting a ggplot with geom_point
}
\details{
Can plot anything possible for the combination ggplot + geom_point. Returns the plot with the generalized
frame of ototools.
}
\examples{
#rm(list = ls())
plot_geom_point(data = gathering, x_var = "oto_diameter", y_var = "tl_mm")
plot_geom_point(data = gathering, x_var = "oto_weight_mg", y_var = "dry_weight_g")

}

