% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Fish growth}
\usage{
plot_growth(data, y_var = "ring_width")
}
\arguments{
\item{data}{a dfr given by master_func}

\item{y_var}{y variable for a geom_line plot}
}
\description{
Plotting ring width
}
\details{
Plots ring width ~ julian day, should not include more than 12 fish and works with the output
from comb_func
}
\examples{
#rm(list = ls())
fish <- "example"
format <- "csv"
data <- master_func(fish = fish, format = format, sep = ";", basic = gathering, catch_day = "catch_date")
plot_growth(data = data)

}

