% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hatch.R
\name{plot_hatch}
\alias{plot_hatch}
\title{plot_hatch}
\usage{
plot_hatch(values, xlab = "")
}
\arguments{
\item{values}{days of hatch}

\item{xlab}{titel of xaxis, default is Hatch_day}
}
\description{
Creating a ggplot with geom_bar.
}
\details{
Meant to create a histogram of the hatch dates,
but simply creates a histogram with the ototools specific frame.
}
\examples{
#rm(list = ls())
values <- basic$deathday - basic$age
plot_hatch(values = values)

}
