% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outline.R
\name{plot_outline}
\alias{plot_outline}
\title{Plot outline}
\usage{
plot_outline()
}
\description{
Gives a frame for plots
}
\details{
Specifies the way plots produced by ototools will look like
}
\examples{
#rm(list = ls())

}

