% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_rows.R
\name{sum_rows}
\alias{sum_rows}
\title{sum_rows}
\usage{
sum_rows(x)
}
\arguments{
\item{x}{a vector containing values that should be summarised.}
}
\description{
Summarise each row with the rows before
}
\details{
Initially meant to calculate the otolith radius at increment x.
If you need it otherwise, go for it.
}
\examples{
#rm(list = ls())
x <- dat$ring_width[[1]]

}
