% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_oto.R
\name{test_oto}
\alias{test_oto}
\title{test_oto}
\usage{
test_oto(fish, format = "txt", sep = "\\t", dir = getwd(),
  section = NULL)
}
\arguments{
\item{fish}{name of the file to be loaded}

\item{format}{format of the file to be loaded (either txt or csv); default is txt}

\item{sep}{the separator used by the file to be loaded; default is tab}

\item{dir}{working directory}

\item{section}{which sections to be loaded. Has to be part of the filename!}
}
\description{
Quick check for one otolith.
}
\details{
Uses load_fish + oto_makro + plot_growth. Returns a dfr and a plot.
}
\examples{
#rm(list = ls())
fish <- "708"
format <- "csv"
test_oto(fish = fish, format = format, sep = ";", dir = "G:/ototools")

}
