% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_func.R
\name{comb_func}
\alias{comb_func}
\title{comb_data}
\usage{
comb_func(x, y, by = "id")
}
\arguments{
\item{x}{a tibble given by oto_makro}

\item{y}{additional informations for (each) fish in x}

\item{by}{columns to be joined by.}
}
\description{
Combine two tables.
}
\details{
Performs a left_join with two datasets. Will merge by id as default, but can merge by every other
column as well.
}
\examples{
#rm(list = ls())
x <- oto_makro(x = example[,3], y = example[,4], fish_no = "708")
y <- basic
y$age <- as.numeric(y$age)
combi <- comb_func(x = x, y = y, by = c("id", "age"))

}
