% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{example}
\format{A data frame with 123 rows and 18 variables:
\describe{
  \item{\code{V1}}{integer. increasing numbers from 1 on.}
  \item{\code{V2}}{integer. same as X.}
  \item{\code{V3}}{integer. x coordinates on the foto.}
  \item{\code{V4}}{integer. y coordinates on the foto.}
  \item{\code{V5}}{integer. not relevant.}
  \item{\code{V6}}{integer. not relevant.}
  \item{\code{V7}}{integer. not relevant.}
  \item{\code{V8}}{integer. not relevant.}
  \item{\code{V9}}{integer. not relevant.}
  \item{\code{V10}}{integer. not relevant.}
  \item{\code{V11}}{integer. not relevant.}
  \item{\code{V12}}{integer. not relevant.}
  \item{\code{V13}}{integer. not relevant.}
  \item{\code{V14}}{integer. not relevant.}
  \item{\code{V15}}{integer. not relevant.}
  \item{\code{V16}}{integer. not relevant.}
  \item{\code{V17}}{integer. not relevant.}
  \item{\code{V18}}{integer. not relevant.}
}}
\usage{
example
}
\description{
exel output generated by imagePRO for fish 708
}
\keyword{datasets}
