% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{example}
\alias{example}
\title{dummy data}
\format{Dummy data:
\describe{
\item{x pos}{x position}
\item{y pos}{y position}
}}
\usage{
example
}
\description{
dummy data
}
\keyword{datasets}

