% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fish.R
\name{load_fish}
\alias{load_fish}
\title{Finding function}
\usage{
load_fish(fish, format = "txt", sep = ";")
}
\arguments{
\item{fish}{name of the file to be loaded}

\item{format}{format of the file to be loaded (either txt or csv, the default is txt)}

\item{sep}{separator used by the file to be loaded (default is ";")}
}
\description{
Loads the data correctly for the makro
}
\details{
Automatically load in the data the way it's needed to work with oto_makro
}
\examples{
#rm(list = ls())
fish <- "example"
format <- "csv"
fish <- load_fish(fish = fish, format = format)
str(fish)

}

