% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{makro_output}
\alias{makro_output}
\title{makro output}
\format{A data frame with 1 rows and 6 variables:
\describe{
  \item{\code{id}}{character. id}
  \item{\code{ring_no}}{list. number of rings.}
  \item{\code{ring_width}}{list. ring width in µm.}
  \item{\code{hc}}{double. hatchcheck in µm.}
  \item{\code{oto_rad}}{double. radius of the otolith in µm.}
  \item{\code{age}}{double. age of the fish in days.}
}}
\usage{
makro_output
}
\description{
output as created by ototools
}
\keyword{datasets}
