% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_func.R
\name{master_func}
\alias{master_func}
\title{master_func}
\usage{
master_func(fish, format = "txt", sep = "\\t", basic = NULL,
  catch_day = NULL, dir = getwd(), section = NULL)
}
\arguments{
\item{fish}{name of the files to be loaded}

\item{format}{format of the files to be loaded (either txt or csv);
has to be the same for every file, default is txt}

\item{sep}{the separator used by the files to be loaded;
has to be the same for every file, default is tab}

\item{basic}{a dfr with some basic measurements for the fish analysed}

\item{catch_day}{name of the column containing the catch day in julian days}

\item{dir}{path to a folder common to all elements of fish as well as basic}

\item{section}{which sections to be loaded. Has to be part of the filename!}
}
\description{
Runs some functions of ototools.
}
\details{
Uses load_fish + oto_makro + master_tab for a list with n filenames.
If basic is given, it will merge basic to the mastertable created by master_tab.
If catch_day is given it will add julday in addition to ring_no.
Will automatically add the moving average based on the method published by Huessy et al. 2003.
Will automatically calculate the otolith radius at increment x.
}
\examples{
#rm(list = ls())
fish <- c("708", "708")
format <- "csv"
sep <- ";"
test <- basic[, names(basic) != "age"]
catch_day <- "deathday"

x <- master_func(fish = fish, format = format, sep = sep,
                 basic = test, catch_day = catch_day, dir = substr(getwd(), 1, 11))
x

x <- master_func(fish = fish, format = format, sep = sep, dir = substr(getwd(), 1, 11))
x

}
