% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_func.R
\name{master_func}
\alias{master_func}
\title{Master function}
\usage{
master_func(fish, format = "txt", sep = "\\t", basic = NULL,
  catch_day = NULL, dir = getwd())
}
\arguments{
\item{fish}{name of the files to be loaded}

\item{format}{format of the files to be loaded (either txt or csv);
has to be the same for every file, default is txt}

\item{sep}{the separator used by the files to be loaded;
has to be the same for every file, default is tab}

\item{basic}{a dfr with some basic values for the fish}

\item{catch_day}{name of the column containing the catch day (julian)}

\item{dir}{path to the folder to be worked in}
}
\description{
Run some functions of ototools
}
\details{
Uses load_fish + oto_makro + master_tab. If basic is given, it will merge basic to the mastertable
with comb_func. If catch_day is given it will add julday and moving average. Returns a dfr.
}
\examples{
#rm(list = ls())
fish <- c("example", "example")
format <- "csv"
sep <- ";"
basic <- gathering[, names(gathering) != "age"]
catch_day <- "catch_date"

x <- master_func(fish = fish, format = format, sep = sep,
                 basic = gathering, catch_day = "catch_date", dir = "F:/ototools")
head(x)

x <- master_func(fish = fish, format = format, sep = sep, dir = "F:/ototools")
head(x)

}
