% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_tab.R
\name{master_tab}
\alias{master_tab}
\title{master_tab}
\usage{
master_tab(x)
}
\arguments{
\item{x}{list with dfrs/ tibbles / ...}
}
\description{
Merges objects of a list.
}
\details{
Uses rbind() to connect all elements of a given list in one object.
Works only with data arranged in columns.
}
\examples{
#rm(list = ls())
x <- list(x = makro_output, x = makro_output)
master_tab(x)

}
