% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_tab.R
\name{master_tab}
\alias{master_tab}
\title{master_tab}
\usage{
master_tab(x)
}
\arguments{
\item{x}{list of dfrs}
}
\description{
Merges dfrs
}
\details{
Does a rbind() for a list of dfrs. Returns the final dfr.
}
\examples{
#rm(list = ls())
x <- list(x = "a", x = "a")
master_tab(x)

}

