% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_ave.R
\name{move_ave}
\alias{move_ave}
\title{Move_ave}
\usage{
move_ave(ring_no, ring_width)
}
\arguments{
\item{ring_no}{column with ring numbers}

\item{ring_width}{column with ring widths}
}
\description{
Calculate the difference in moving averages following Huessy et al. 2003
}
\details{
Calculating the difference of the moving average from the last 10 increments and the next
10 increments. Returns absolut values.
}
\examples{
#rm(list = ls())
dat <- oto_makro(x = example[,3], y = example[,4], fish_no = 1)
ring_no <- dat$ring_no[[1]]
ring_width <- dat$ring_width[[1]]
move_ave(ring_no, ring_width)
}
