% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_makro.R
\name{oto_makro}
\alias{oto_makro}
\title{Otolith reading Makro}
\usage{
oto_makro(x_coord, y_coord, fish_no)
}
\arguments{
\item{x_coord}{a vector that will be converted to numeric, is supposed to be x coordinates
from imagePRO}

\item{y_coord}{a vector that will be converted to numeric, is supposed to be x coordinates
from imagePRO}

\item{fish_no}{number of the fish or whatever chosen to identify them}
}
\description{
Function to transform coordinates from imagePRO into distances
}
\details{
Function was created to deal with the output from imagePRO. It simply applies the
Pythagoras' theorem between each line and the folling one.
Returns a dfr with the folloing values: ring_no, ring_width, hc, oto_rad, age, id
}
\examples{
#rm(list = ls())
#returns a data.frame with ring_width and more
x <- oto_makro(x = example[,3], y = example[,4], fish_no = 1)
head(x)

}

