% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_makro.R
\name{oto_makro}
\alias{oto_makro}
\title{oto_makro}
\usage{
oto_makro(x_coord, y_coord, fish_no)
}
\arguments{
\item{x_coord}{a vector that will be converted to numeric, is supposed to be x coordinates
from imagePRO}

\item{y_coord}{a vector that will be converted to numeric, is supposed to be y coordinates
from imagePRO}

\item{fish_no}{number of the fish or whatever chosen to identify different individual}
}
\description{
Otolith reading makro.
}
\details{
The function was created to deal with the output from imagePRO. It simply applies the
Pythagoras' theorem between each line and the folling one of two vectors.
Returns a tibble with the folloing values: ring_no, ring_width, hc, oto_rad, age and id.
}
\examples{
#rm(list = ls())
oto_makro(x_coord = example[,3], y_coord = example[,4], fish_no = 1)

}
