% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geom_point.R
\name{plot_geom_point}
\alias{plot_geom_point}
\title{plot_geom_point}
\usage{
plot_geom_point(data, x_var, y_var)
}
\arguments{
\item{data}{a dfr holding x_var and y_var}

\item{x_var}{x variable for a geom_point plot}

\item{y_var}{y variable for a geom_point plot}
}
\description{
Creates a ggplot with geom_point.
}
\details{
Meant to create plots to display different correlations between measurements taken to evaluate the
growth of fish, but simply creates a geom_point based plot with the ototools specific frame.
}
\examples{
#rm(list = ls())
plot_geom_point(data = basic, x_var = "oto_diameter_mm", y_var = "tl_mm")
plot_geom_point(data = basic, x_var = "oto_weight_mg", y_var = "dry_weight_g")

}
