% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Fish growth}
\usage{
plot_growth(data, catch_day, y_var = "ring_width")
}
\arguments{
\item{data}{a dfr given by comb_func}

\item{catch_day}{name of the column containing the catch day (julian)}
}
\description{
Plotting ring width
}
\details{
Plots ring width ~ julian day, should not include more than 12 fish and works with the output
from comb_func
}
\examples{
#rm(list = ls())
x <- oto_makro(x = example[,3], y = example[,4], fish_no = "example")$values
y <- gathering
data <- comb_func(x = x, y = y)
catch_day <- "catch_date"
plot_growth(data = data, catch_day = catch_day)

}

