% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{plot_growth}
\usage{
plot_growth(x_var, y_var)
}
\arguments{
\item{x_var}{x variable for a geom_line plot}

\item{y_var}{y variable for a geom_line plot}
}
\description{
Creates a ggplot with geom_line.
}
\details{
Meant to plot the growth of a fish based on otlith analysis,
but simply creates a geom_line based plot with the ototools specific frame.
}
\examples{
#rm(list = ls())
x_var <- dat$julday
y_var <- dat$ring_width
p <- purrr::map2(x_var, y_var, ~plot_growth(x_var = .x, y_var = .y))
p[[1]] + ggplot2::labs(x = "Julday", y = "Ring_width")
\dontrun{
purrr::map2(p, list(c(1,2)), ~.x + ggplot2::geom_hline(yintercept = .y))
}

}
