% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Fish growth}
\usage{
plot_growth(x_var, y_var)
}
\arguments{
\item{x_var}{x variable for a geom_line plot}

\item{y_var}{y variable for a geom_line plot}
}
\description{
Plotting ring width ~ julday on default
}
\details{
Creates a ggplot with geom_line, should not include more than 12 fish if id is TRUE.
Can be used to create other geom_line plots. Returns the plot.
}
\examples{
#rm(list = ls())
fish <- c("example", "example")
format <- "csv"
x_var <- master_tib$julday
y_var <- master_tib$ring_width
p <- purrr::map2(x_var, y_var, ~plot_growth(x_var = .x, y_var = .y))
p[[1]] + ggplot2::labs(x = "Julday", y = "Ring_width")
purrr::map2(p, c(1,2), ~.x + ggplot2::geom_hline(yintercept = .y))

}
