% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Fish growth}
\usage{
plot_growth(data, x_var = "julday", y_var = "ring_width", id = TRUE)
}
\arguments{
\item{data}{a dfr holding both x_var and y_var}

\item{x_var}{x variable for a geom_line plot}

\item{y_var}{y variable for a geom_line plot}

\item{id}{if id is TRUE one plot for each id will be produced,
if not plot_growth will only return one plot}
}
\description{
Plotting ring width
}
\details{
Plots ring width ~ julian day, should not include more than 12 fish and works with the output
from comb_func. Can be used to create other geom_line plots
}
\examples{
#rm(list = ls())
fish <- "example"
format <- "csv"
data <- master_func(fish = fish, format = format, sep = ";",
                    basic = gathering, catch_day = "catch_date")
plot_growth(data = data)

}

