% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hatch.R
\name{plot_hatch}
\alias{plot_hatch}
\title{Hatch histogram}
\usage{
plot_hatch(values, xlab = "Hatch_day")
}
\arguments{
\item{values}{days of hatch}

\item{xlab}{titel of xaxis, default is Hatch_day}
}
\description{
Creating a histogram
}
\details{
Creating a histogram of the hatch dates. Values should be a vector with julian days.
Returns the plot.
}
\examples{
#rm(list = ls())
values <- gathering$catch_date - gathering$age
plot_hatch(values = values)

}

