% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hatch.R
\name{plot_hatch}
\alias{plot_hatch}
\title{Hatch histogram}
\usage{
plot_hatch(values, xlab = "Hatch_day")
}
\arguments{
\item{values}{days of hatch in julian days}

\item{xlab}{Titel of xaxis, default is Hatch_day}
}
\description{
Creating a histogram
}
\details{
Creating a histogram of the hatch dates
}
\examples{
#rm(list = ls())
gathering$hatch_day <- gathering$catch_date - gathering$age
values <- gathering$hatch_day
plot_hatch(values = values)

}

