% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{test data}
\format{Dummy data:
\describe{
\item{x pos}{x position}
\item{y pos}{y position}
}}
\usage{
test
}
\description{
result from the exel makro for "example"
}
\keyword{datasets}
